Remove InPageNavigation 'simple' variant
