+++ content/webapp/views/pages/about-us/cookie-policy.tsx
@@ -77,7 +77,6 @@ const CookiePolicyPage: NextPage<page.Props> = props => {
           <Container>
             <Grid style={{ background: 'white', rowGap: 0 }}>
               <InPageNavigation
-                variant="sticky"
                 links={onThisPage!}
                 sizeMap={{ s: [12], m: [12], l: [3], xl: [3] }}
                 isOnWhite
